+++ youtube_dl.go
@@ -53,10 +53,14 @@ type AudioPlaylist struct {
 // Download downloads the song via youtube-dl if it does not already exist on disk.
 // All downloaded songs are stored in ~/.mumbledj/songs and should be automatically cleaned.
 func (dl *AudioTrack) Download() error {
+	player := "--prefer-ffmpeg"
+	if dj.conf.General.PlayerCommand == "avconv" {
+		player = "--prefer-avconv"
+	}
 
 	// Checks to see if song is already downloaded
 	if _, err := os.Stat(fmt.Sprintf("%s/.mumbledj/songs/%s", dj.homeDir, dl.Filename())); os.IsNotExist(err) {
-		cmd := exec.Command("youtube-dl", "--verbose", "--no-mtime", "--output", fmt.Sprintf("%s/.mumbledj/songs/%s", dj.homeDir, dl.Filename()), "--format", dl.format, "--prefer-ffmpeg", dl.url)
+		cmd := exec.Command("youtube-dl", "--verbose", "--no-mtime", "--output", fmt.Sprintf("%s/.mumbledj/songs/%s", dj.homeDir, dl.Filename()), "--format", dl.format, player, dl.url)
 		output, err := cmd.CombinedOutput()
 		if err == nil {
 			if dj.conf.Cache.Enabled {